requirements.txt
Pillow
noise
